[BE] Controller Test (#329)
* refactor: package 구조 변경

* test: ArchiveController Test

* test: ColorController Test

* test: EstimationController Test

* test: OptionController Test

* test: TrimController Test

* test: UserController Test

* refactor: 사용하지 않는 클래스 제거

* test: BookmarkRepository - 존재하지 않는 ID로 Bookmark를 조회했을 때 테스트

* CarArchivingRepository - EmptyResultDataAccessException에 대한 테스트

* test: CarOptionRepositoryTest - result가 Empty일 때 테스트

* test: CarReviewRepository - EmptyResultDataAccessException에 대한 테스트

* test: UserRepository Test

* test: 내가 만든 차량 혹은 피드에서 저장한 차량이 없을 때 테스트

* test: UserService Test

* fix: access-token, refresh-token expiration 수정
