--- backend/src/main/java/com/backend/topcariving/domain/dto/option/request/esitmation/ModifiedRequestDTO.java
@@ -1,13 +1,0 @@
-package com.backend.topcariving.domain.dto.option.request.esitmation;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class ModifiedRequestDTO {
-
-	private Long archivingId;
-
-	private Long carOptionId;
-}
