--- backend/src/main/java/com/backend/topcariving/domain/dto/projection/tag/TagRank.java
@@ -1,20 +1,0 @@
-package com.backend.topcariving.domain.dto.projection.tag;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.ToString;
-
-@AllArgsConstructor
-@Getter
-@ToString
-public class TagRank {
-
-	private Long carOptionId;
-
-	private String tagText;
-
-	private String optionName;
-
-	private Integer reviewCount;
-
-}
